feat: update favorites functionality and restructure components
- Replaced the `Favorites` component with `FavoritesView` for improved structure and user experience.
- Updated import paths for favorite actions to align with the new directory structure under `features/favorites`.
- Introduced a new `useFavoriteActions` hook to manage adding and removing favorites with toast notifications.
- Enhanced the `FavoritesView` to display favorite items and allow users to add them to the cart or remove them from favorites.
