+++ components/shared/sections/Collection.tsx
@@ -4,7 +4,7 @@ import ProductCard from "@/components/shared/ProductCard";
 import Title from "@/components/shared/Title";
 import { useProduct } from "@/features/admin/hooks/useProduct";
 import { useCartActions } from "@/features/cart/hooks/useCartActions";
-import { useFavoriteActions } from "@/hooks/useFavoriteActions";
+import { useFavoriteActions } from "@/features/favorites/hooks/useFavoriteActions";
 import { useFavoriteStore } from "@/stores/FavoriteStore";
 
 export default function Collection() {
